i think it works?
1) use actual coords of tile as bounding box to clamp coords of
downloaded orthophoto. in fsx/p3d, it seems that it doesn't matter to
use coords of orthophoto. the remainder is just masked out by the black
of the mask. in fs2004, if you don't use the actual coordinates of the
tile, bad things happen because then you get a "placeholder" of just
water that is the size of the orthophoto coordinates but the actual
masked out orthophoto (corresponding to the size of the tile) is the
incorrect size to fir this placeholder.
2) also, use round instead of ceil and floor to clip the u and v lod13
cell calculation