crop unused parts of mask and main image
needed for fs9 resample as it can't show underlying terrain or water.
all black just shows water for fs9. leave it for fsx/p3d so we don't
resample extra for no reason

Might eventually do it on creation of mask and main ortho image, but
not sure as seems like a lot of work